+++ AdminWebsite/AdminWebsite/ClientApp/src/app/booking/hearing-schedule/hearing-schedule.component.html
@@ -136,7 +136,7 @@
         </div>
       </div>
     </div>
-    <div *ngIf="hearing.isMultiDayEdit && form.get('dates').controls.length > 0">
+    <div *ngIf="hearing.isMultiDayEdit && form.get('newDates').controls.length > 0">
       <table class="govuk-table govuk-!-width-one-half" id="hearings-in-group-dates-table" aria-describedby="multi day hearing dates">
         <thead class="govuk-table__head">
           <tr class="govuk-table__row">
@@ -145,7 +145,7 @@
           </tr>
         </thead>
         <tbody class="govuk-table__body">
-          <tr class="govuk-table__row" *ngFor="let dateControl of form.get('dates').controls; let i = index">
+          <tr class="govuk-table__row" *ngFor="let dateControl of form.get('newDates').controls; let i = index">
             <td class="govuk-table__cell">{{ formatDate(hearingsInGroupToEdit[i]?.originalScheduledDateTime) }}</td>
             <td class="govuk-table__cell" [ngClass]="{ 'govuk-form-group--error': newDatesInvalid }">
               <input class="govuk-input" type="date" [formControl]="dateControl" min="{{ today | date : 'yyyy-MM-dd' }}" />
